Create controller file
